Addressing PR comments

+++ mlir/test/Dialect/Rock/affix_tuning_params.mlir
@@ -643,7 +643,7 @@ func.func @rock_conv_tuning(%arg0: tensor<1x1x1x3x3xf32>, %arg1: tensor<64x1x1x1
 }
 
 // CHECK-LABEL: @rock_attn_numstages2
-func.func @rock_attn_numstages2(%arg0: tensor<32768xf16>, %arg1: tensor<32768xf16>, %arg2: tensor<32768xf16>, %arg3: tensor<32768xf16>) -> tensor<32768xf16> attributes {rock.arch = "amdgcn-amd-amdhsa:gfx1100", rock.kernel, rock.num_stages = 2 : i32} {
+func.func @rock_attn_numstages2(%arg0: tensor<32768xf16>, %arg1: tensor<32768xf16>, %arg2: tensor<32768xf16>, %arg3: tensor<32768xf16>) -> tensor<32768xf16> attributes {rock.arch = "amdgcn-amd-amdhsa:gfx1100", rock.kernel, rock.num_stages = 2 : i64} {
   %0 = rock.transform %arg0 by <affine_map<(d0, d1, d2) -> (d1 * 32 + d2)> by [<Unmerge{1024, 32} ["seq_q", "head_qk"] at [1, 2] -> ["raw"] at [0]>, <AddDim{1} ["g"] at [0] -> [] at []>] bounds = [1, 1024, 32] -> [32768]> : tensor<32768xf16> to tensor<1x1024x32xf16>
   %1 = rock.transform %arg1 by <affine_map<(d0, d1, d2) -> (d1 * 1024 + d2)> by [<Unmerge{32, 1024} ["head_qk", "seq_k"] at [1, 2] -> ["raw"] at [0]>, <AddDim{1} ["g"] at [0] -> [] at []>] bounds = [1, 32, 1024] -> [32768]> : tensor<32768xf16> to tensor<1x32x1024xf16>
   %2 = rock.transform %arg2 by <affine_map<(d0, d1, d2) -> (d1 * 32 + d2)> by [<Unmerge{1024, 32} ["seq_k", "head_v"] at [1, 2] -> ["raw"] at [0]>, <AddDim{1} ["g"] at [0] -> [] at []>] bounds = [1, 1024, 32] -> [32768]> : tensor<32768xf16> to tensor<1x1024x32xf16>
@@ -663,7 +663,7 @@ func.func @rock_attn_numstages2(%arg0: tensor<32768xf16>, %arg1: tensor<32768xf1
 }
 
 // CHECK-LABEL: @rock_attn_numstages3
-func.func @rock_attn_numstages3(%arg0: tensor<32768xf16>, %arg1: tensor<32768xf16>, %arg2: tensor<32768xf16>, %arg3: tensor<32768xf16>) -> tensor<32768xf16> attributes {rock.arch = "amdgcn-amd-amdhsa:gfx950", rock.kernel, rock.num_stages = 3 : i32} {
+func.func @rock_attn_numstages3(%arg0: tensor<32768xf16>, %arg1: tensor<32768xf16>, %arg2: tensor<32768xf16>, %arg3: tensor<32768xf16>) -> tensor<32768xf16> attributes {rock.arch = "amdgcn-amd-amdhsa:gfx950", rock.kernel, rock.num_stages = 3 : i64} {
   %0 = rock.transform %arg0 by <affine_map<(d0, d1, d2) -> (d1 * 32 + d2)> by [<Unmerge{1024, 32} ["seq_q", "head_qk"] at [1, 2] -> ["raw"] at [0]>, <AddDim{1} ["g"] at [0] -> [] at []>] bounds = [1, 1024, 32] -> [32768]> : tensor<32768xf16> to tensor<1x1024x32xf16>
   %1 = rock.transform %arg1 by <affine_map<(d0, d1, d2) -> (d1 * 1024 + d2)> by [<Unmerge{32, 1024} ["head_qk", "seq_k"] at [1, 2] -> ["raw"] at [0]>, <AddDim{1} ["g"] at [0] -> [] at []>] bounds = [1, 32, 1024] -> [32768]> : tensor<32768xf16> to tensor<1x32x1024xf16>
   %2 = rock.transform %arg2 by <affine_map<(d0, d1, d2) -> (d1 * 32 + d2)> by [<Unmerge{1024, 32} ["seq_k", "head_v"] at [1, 2] -> ["raw"] at [0]>, <AddDim{1} ["g"] at [0] -> [] at []>] bounds = [1, 1024, 32] -> [32768]> : tensor<32768xf16> to tensor<1x1024x32xf16>
@@ -683,7 +683,7 @@ func.func @rock_attn_numstages3(%arg0: tensor<32768xf16>, %arg1: tensor<32768xf1
 }
 
 // CHECK-LABEL: @rock_attn_numstages4
-func.func @rock_attn_numstages4(%arg0: tensor<32768xf16>, %arg1: tensor<32768xf16>, %arg2: tensor<32768xf16>, %arg3: tensor<32768xf16>) -> tensor<32768xf16> attributes {rock.arch = "amdgcn-amd-amdhsa:gfx950", rock.kernel, rock.num_stages = 4 : i32} {
+func.func @rock_attn_numstages4(%arg0: tensor<32768xf16>, %arg1: tensor<32768xf16>, %arg2: tensor<32768xf16>, %arg3: tensor<32768xf16>) -> tensor<32768xf16> attributes {rock.arch = "amdgcn-amd-amdhsa:gfx950", rock.kernel, rock.num_stages = 4 : i64} {
   %0 = rock.transform %arg0 by <affine_map<(d0, d1, d2) -> (d1 * 32 + d2)> by [<Unmerge{1024, 32} ["seq_q", "head_qk"] at [1, 2] -> ["raw"] at [0]>, <AddDim{1} ["g"] at [0] -> [] at []>] bounds = [1, 1024, 32] -> [32768]> : tensor<32768xf16> to tensor<1x1024x32xf16>
   %1 = rock.transform %arg1 by <affine_map<(d0, d1, d2) -> (d1 * 1024 + d2)> by [<Unmerge{32, 1024} ["head_qk", "seq_k"] at [1, 2] -> ["raw"] at [0]>, <AddDim{1} ["g"] at [0] -> [] at []>] bounds = [1, 32, 1024] -> [32768]> : tensor<32768xf16> to tensor<1x32x1024xf16>
   %2 = rock.transform %arg2 by <affine_map<(d0, d1, d2) -> (d1 * 32 + d2)> by [<Unmerge{1024, 32} ["seq_k", "head_v"] at [1, 2] -> ["raw"] at [0]>, <AddDim{1} ["g"] at [0] -> [] at []>] bounds = [1, 1024, 32] -> [32768]> : tensor<32768xf16> to tensor<1x1024x32xf16>
